put the fuse before the reverse voltage protection (thanks to bktemp from eevblog forum)
Quoting bktemp user on eevblog:

When the fuse fails, there will be a voltage spike.
Since there is no load except the mosfet gate, the voltage spike will damage the mosfet,
maybe even shorting gate-source and therefore the input supply.
If the gate is connected after the fuse, is will be connected in parallel with C1,
providing some protection for the sensitive gate.